Simple input check/usage for core. Unique tmp file
Now there is a unique tmp file for every call of
ZPriMATE to enable simultaneous running of different
instances of the program. To this end the path of the tmp
file was added as an additional input parameter for main.cpp/core.

Simple check implemented if proper number of arguments for core
are given, file existance, etc...

Not yet tested on MAC